refactor: inter-protocol/src/vaultFactory/types-ambient.js
