+++ packages/inter-protocol/exported.js
@@ -1,2 +1,2 @@
 import './src/psm/types.js';
-import './src/vaultFactory/types.js';
+import './src/vaultFactory/types-ambient.js';
