+++ packages/inter-protocol/src/vaultFactory/type-imports.js
@@ -1,4 +1,4 @@
 // @jessie-check
 
 import '@agoric/zoe/exported.js';
-import './types.js';
+import './types-ambient.js';
