+++ packages/inter-protocol/test/vaultFactory/driver.js
@@ -21,7 +21,7 @@ import {
   startVaultFactory,
 } from '../../src/proposals/econ-behaviors.js';
 import { startEconomicCommittee } from '../../src/proposals/startEconCommittee.js';
-import '../../src/vaultFactory/types.js';
+import '../../src/vaultFactory/types-ambient.js';
 import {
   installPuppetGovernance,
   setupBootstrap,
