+++ packages/inter-protocol/test/vaultFactory/test-storage.js
@@ -7,7 +7,7 @@ import { E } from '@endo/eventual-send';
 import { assertTopicPathData, subscriptionKey } from '../supports.js';
 import { makeDriverContext, makeManagerDriver } from './driver.js';
 
-import '../../src/vaultFactory/types.js';
+import '../../src/vaultFactory/types-ambient.js';
 
 /** @typedef {import('./driver.js').DriverContext & {}} Context */
 /** @type {import('ava').TestFn<Context>} */
