+++ packages/inter-protocol/test/vaultFactory/test-vaultFactory.js
@@ -26,7 +26,7 @@ import { deeplyFulfilled } from '@endo/marshal';
 import { calculateCurrentDebt } from '../../src/interest-math.js';
 import { SECONDS_PER_YEAR } from '../../src/interest.js';
 import { startVaultFactory } from '../../src/proposals/econ-behaviors.js';
-import '../../src/vaultFactory/types.js';
+import '../../src/vaultFactory/types-ambient.js';
 import {
   metricsTracker,
   reserveInitialState,
