+++ packages/inter-protocol/test/vaultFactory/vaultFactoryUtils.js
@@ -12,7 +12,7 @@ import {
   setupReserve,
   startAuctioneer,
 } from '../../src/proposals/econ-behaviors.js';
-import '../../src/vaultFactory/types.js';
+import '../../src/vaultFactory/types-ambient.js';
 import {
   installPuppetGovernance,
   produceInstallations,
